feat(various): Places the Agent and User nav links into corresponding NavDropdowns
Removes the useNavLinks hooks; Moves the NavLinkConfig type into the CustomNavLink component; Now directly constructing the dropdown items for the profile and sign out links
