+++ src/features/navbar/index.ts
@@ -1,2 +1,1 @@
 export * from './components/BitwiseNavbar';
-export * from './hooks/useNavLinks';
